+++ src/lib/rxfm/components/chainable-creator.test.ts
@@ -4,7 +4,7 @@ import { Div, Button, Input } from "./html";
 
 const testComponent = <T extends ElementType>(component: Component<T>) => {
   let element: T | undefined = undefined;
-  const subscription = component.subscribe(el => { element = el });
+  const subscription = component.subscribe(el => { element = el; });
   return { element: element!, unsubscribe: () => subscription.unsubscribe() };
 };